refactor: extract env loading for tracing
`setup_tracing` requires a few environment variables that can be
refactored into a standalone struct and accompanied function that can
load the required variables as well as provide sane defaults instead of
simply panicking when they are not found.

`crate_name` can be excluded as it is provided by Cargo.

Signed-off-by: Ivin Joel Abraham <ivinjoelabraham@proton.me>