refactor: introduce reusable reqwest client
The primary motivation for this change is to avoid unnecessarily
repeated calls to the env. var. ROOT_URL and more importantly, creating
new reqwest clients for each query and mutation. This has also resulted
in centralizing the ROOT_URL variable at the cost of introducing a new
struct to store in Data.

Signed-off-by: Ivin Joel Abraham <ivinjoelabraham@proton.me>